Remove namespace
