+++ sources/Google.Solutions.Settings.Test/RegistrySettingsStoreBase.cs
@@ -23,7 +23,7 @@ using Google.Solutions.Settings.Registry;
 using Microsoft.Win32;
 using NUnit.Framework;
 
-namespace Google.Solutions.Settings.Test.Registry
+namespace Google.Solutions.Settings.Test
 {
     public abstract class RegistrySettingsStoreBase
     {
