+++ sources/Google.Solutions.Settings/DictionarySettingsStore.cs
