+++ sources/Google.Solutions.Settings/DictionaryValueAccessor.cs
