+++ sources/Google.Solutions.Settings/ISettingsStore.cs
