+++ sources/Google.Solutions.Settings/IValueAccessor.cs
