+++ sources/Google.Solutions.Settings/MergedSettingsStore.cs
