+++ sources/Google.Solutions.Settings/PolicyEnabledRegistryRepository.cs
