+++ sources/Google.Solutions.Settings/RegistryRepositoryBase.cs
