+++ sources/Google.Solutions.Settings/RegistrySetting.cs
