+++ sources/Google.Solutions.Settings/RegistrySettingsStore.cs
