+++ sources/Google.Solutions.Settings/RegistryValueAccessor.cs
