+++ sources/Google.Solutions.Settings/SettingsStoreBase.cs
